feat: add code running functionality with the Piston API
This update adds a new command that allows users to run code in a variety of languages using the Piston API. The languages and their versions are pulled from a new JSON file. An accompanying modal also provides options for input parameters. Autocomplete functionality for language and version selection is included.

A feature to fix bugs with AI is also included but may be janky. Need to work on this